mario walk ground
지형만 충돌 확인

+++ finalProject/finalProject/boss.h
@@ -52,7 +52,7 @@ public:
 		targetPoint = playerPoint;
 	}
 
-	void moveX(int X)
+	void moveX(float X)
 	{
 		p.x += X;
 	}
